+++ howso/direct/core.py
@@ -727,342 +727,6 @@ class HowsoCore:
             "impute_session": impute_session,
         })
 
-    def react(
-        self,
-        trainee_id: str,
-        *,
-        action_features: Optional[Iterable[str]] = None,
-        action_values: Optional[List[List[object]]] = None,
-        allow_nulls: bool = False,
-        case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        context_features: Optional[Iterable[str]] = None,
-        context_values: Optional[List[List[object]]] = None,
-        derived_action_features: Optional[Iterable[str]] = None,
-        derived_context_features: Optional[Iterable[str]] = None,
-        desired_conviction: Optional[float] = None,
-        details: Optional[Dict] = None,
-        exclude_novel_nominals_from_uniqueness_check: bool = False,
-        extra_features: Optional[Iterable[str]] = None,
-        feature_bounds_map: Optional[Dict] = None,
-        generate_new_cases: Literal["always", "attempt", "no"] = "no",
-        input_is_substituted: bool = False,
-        into_series_store: Optional[str] = None,
-        leave_case_out: bool = False,
-        new_case_threshold: Literal["max", "min", "most_similar"] = "min",
-        ordered_by_specified_features: bool = False,
-        post_process_features: Optional[Iterable[str]] = None,
-        post_process_values: Optional[List[object]] = None,
-        preserve_feature_values: Optional[Iterable[str]] = None,
-        substitute_output: bool = True,
-        use_case_weights: bool = False,
-        use_regional_model_residuals: bool = True,
-        weight_feature: Optional[str] = None
-    ) -> Dict:
-        """
-        Single case react.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        context_values : list of list of object, optional
-            A 2d list of context values to react to.
-            If None for discriminative react, it is assumed that `session`
-            and `session_id` keys are set in the `details`.
-        action_features : iterable of str, optional
-            An iterable of feature names to treat as action features during
-            react.
-        action_values : list of list of object, optional
-            One or more action values to use for action features.
-            If specified, will only return the specified explanation
-            details for the given actions. (Discriminative reacts only)
-        allow_nulls : bool, default False
-            When true will allow return of null values if there
-            are nulls in the local model for the action features, applicable
-            only to discriminative reacts.
-        context_features : iterable of str, optional
-            An iterable of feature names to treat as context features during
-            react.
-        derived_context_features : iterable of str, optional
-            An iterable of feature names whose values should be computed
-            from the provided context in the specified order. Must be different
-            than context_features.
-        derived_action_features : iterable of str, optional
-            An iterable of feature names whose values should be computed
-            after generation from the generated case prior to output, in the
-            specified order. Must be a subset of action_features.
-        input_is_substituted : bool, default False
-            if True assumes provided categorical (nominal or
-            ordinal) feature values have already been substituted.
-        substitute_output : bool, default True
-            If False, will not substitute categorical feature
-            values. Only applicable if a substitution value map has been set.
-        details : dict, optional
-            If details are specified, the response will
-            contain the requested explanation data along with the reaction.
-        desired_conviction : float
-            If specified will execute a generative react. If not
-            specified will executed a discriminative react. Conviction is the
-            ratio of expected surprisal to generated surprisal for each
-            feature generated, valid values are in the range of zero to
-            infinity.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-        use_case_weights : bool, default False
-            If set to True will scale influence weights by each
-            case's weight_feature weight.
-        case_indices : Iterable of Sequence[Union[str, int]], defaults to None
-            An Iterable of Sequences, of session id and index, where
-            index is the original 0-based index of the case as it was trained
-            into the session. If this case does not exist, discriminative react
-            outputs null, generative react ignores it.
-        post_process_features : iterable of str, optional
-            List of feature names that will be made available during the
-            execution of post_process feature attributes.
-        post_process_values : list of object, optional
-            A 2d list of values corresponding to post_process_features that
-            will be made available during the execution of post_process feature
-            attributes.
-        preserve_feature_values : iterable of str
-            List of features that will preserve their values from
-            the case specified by case_indices, appending and overwriting the
-            specified contexts as necessary.  For generative reacts, if
-            case_indices isn't specified will preserve feature values of a
-            random case.
-        leave_case_out : bool, default False
-            If set to True and specified along with case_indices,
-            each individual react will respectively ignore the corresponding
-            case specified by case_indices by leaving it out.
-        into_series_store : str, optional
-            The name of a series store. If specified, will store an internal
-            record of all react contexts for this session and series to be used
-            later with train series.
-        use_regional_model_residuals : bool
-            If false uses model feature residuals, if True
-            recalculates regional model residuals.
-        feature_bounds_map : dict of dict
-            A mapping of feature names to the bounds for the
-            feature values to be generated in.
-        generate_new_cases : {"always", "attempt", "no"}, default "no"
-            How to generate new cases.
-        ordered_by_specified_features : bool, default False
-            If True order of generated feature values will match
-            the order of specified features.
-        new_case_threshold : {"min", "max", "most_similar"}, optional
-            Distance to determine the privacy cutoff. If None,
-            will default to "min".
-        exclude_novel_nominals_from_uniqueness_check : bool, default False
-            If True, will exclude features which have a subtype defined in their feature
-            attributes from the uniqueness check that happens when ``generate_new_cases``
-            is True. Only applies to generative reacts.
-
-        Returns
-        -------
-        dict
-            The react result including audit details.
-        """
-        return self._execute(trainee_id, "react", {
-            "context_features": context_features,
-            "context_values": context_values,
-            "action_features": action_features,
-            "action_values": action_values,
-            "details": details,
-            "derived_action_features": derived_action_features,
-            "derived_context_features": derived_context_features,
-            "exclude_novel_nominals_from_uniqueness_check": exclude_novel_nominals_from_uniqueness_check,
-            "extra_features": extra_features,
-            "case_indices": case_indices,
-            "allow_nulls": allow_nulls,
-            "input_is_substituted": input_is_substituted,
-            "substitute_output": substitute_output,
-            "weight_feature": weight_feature,
-            "leave_case_out": leave_case_out,
-            "use_case_weights": use_case_weights,
-            "use_regional_model_residuals": use_regional_model_residuals,
-            "desired_conviction": desired_conviction,
-            "feature_bounds_map": feature_bounds_map,
-            "generate_new_cases": generate_new_cases,
-            "ordered_by_specified_features": ordered_by_specified_features,
-            "post_process_features": post_process_features,
-            "post_process_values": post_process_values,
-            "preserve_feature_values": preserve_feature_values,
-            "new_case_threshold": new_case_threshold,
-            "into_series_store": into_series_store,
-        })
-
-    def batch_react(
-        self,
-        trainee_id: str,
-        *,
-        action_features: Optional[Iterable[str]] = None,
-        action_values: Optional[List[List[object]]] = None,
-        allow_nulls: bool = False,
-        case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        context_features: Optional[Iterable[str]] = None,
-        context_values: Optional[List[List[object]]] = None,
-        derived_action_features: Optional[Iterable[str]] = None,
-        derived_context_features: Optional[Iterable[str]] = None,
-        desired_conviction: Optional[float] = None,
-        details: Optional[Dict] = None,
-        exclude_novel_nominals_from_uniqueness_check: bool = False,
-        extra_features: Optional[Iterable[str]] = None,
-        feature_bounds_map: Optional[Dict] = None,
-        generate_new_cases: Literal["always", "attempt", "no"] = "no",
-        input_is_substituted: bool = False,
-        into_series_store: Optional[str] = None,
-        leave_case_out: bool = False,
-        new_case_threshold: Literal["max", "min", "most_similar"] = "min",
-        num_cases_to_generate: Optional[int] = None,
-        ordered_by_specified_features: bool = False,
-        post_process_features: Optional[Iterable[str]] = None,
-        post_process_values: Optional[List[List[object]]] = None,
-        preserve_feature_values: Optional[Iterable[str]] = None,
-        substitute_output: bool = True,
-        use_case_weights: bool = False,
-        use_regional_model_residuals: bool = True,
-        weight_feature: Optional[str] = None
-    ) -> Tuple[Dict, int, int]:
-        """
-        Multiple case react.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        context_values : list of list of object, optional
-            A 2d list of context values to react to.
-            If None for discriminative react, it is assumed that `session`
-            and `session_id` keys are set in the `details`.
-        action_features : iterable of str, optional
-            An iterable of feature names to treat as action features during
-            react.
-        action_values : list of list of object, optional
-            One or more action values to use for action features.
-            If specified, will only return the specified explanation
-            details for the given actions. (Discriminative reacts only)
-        allow_nulls : bool, default False
-            When true will allow return of null values if there
-            are nulls in the local model for the action features, applicable
-            only to discriminative reacts.
-        context_features : iterable of str, optional
-            An iterable of feature names to treat as context features during
-            react.
-        derived_context_features : iterable of str, optional
-            An iterable of feature names whose values should be computed
-            from the provided context in the specified order. Must be different
-            than context_features.
-        derived_action_features : iterable of str, optional
-            An iterable of feature names whose values should be computed
-            after generation from the generated case prior to output, in the
-            specified order. Must be a subset of action_features.
-        exclude_novel_nominals_from_uniqueness_check : bool, default False
-            If True, will exclude features which have a subtype defined in their feature
-            attributes from the uniqueness check that happens when ``generate_new_cases``
-            is True. Only applies to generative reacts.
-        input_is_substituted : bool, default False
-            if True assumes provided categorical (nominal or
-            ordinal) feature values have already been substituted.
-        substitute_output : bool, default True
-            If False, will not substitute categorical feature
-            values. Only applicable if a substitution value map has been set.
-        details : dict, optional
-            If details are specified, the response will
-            contain the requested explanation data along with the reaction.
-        desired_conviction : float
-            If specified will execute a generative react. If not
-            specified will executed a discriminative react. Conviction is the
-            ratio of expected surprisal to generated surprisal for each
-            feature generated, valid values are in the range of zero to
-            infinity.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-        use_case_weights : bool, default False
-            If set to True will scale influence weights by each
-            case's weight_feature weight.
-        case_indices : Iterable of Sequence[Union[str, int]], defaults to None
-            An Iterable of Sequences, of session id and index, where
-            index is the original 0-based index of the case as it was trained
-            into the session. If this case does not exist, discriminative react
-            outputs null, generative react ignores it.
-        post_process_features : iterable of str, optional
-            List of feature names that will be made available during the
-            execution of post_process feature attributes.
-        post_process_values : list of list of object, optional
-            A 2d list of values corresponding to post_process_features that
-            will be made available during the execution of post_process feature
-            attributes.
-        preserve_feature_values : iterable of str
-            List of features that will preserve their values from
-            the case specified by case_indices, appending and overwriting the
-            specified contexts as necessary.  For generative reacts, if
-            case_indices isn't specified will preserve feature values of a
-            random case.
-        leave_case_out : bool, default False
-            If set to True and specified along with case_indices,
-            each individual react will respectively ignore the corresponding
-            case specified by case_indices by leaving it out.
-        into_series_store : str, optional
-            The name of a series store. If specified, will store an internal
-            record of all react contexts for this session and series to be used
-            later with train series.
-        use_regional_model_residuals : bool
-            If false uses model feature residuals, if True
-            recalculates regional model residuals.
-        feature_bounds_map : dict of dict
-            A mapping of feature names to the bounds for the
-            feature values to be generated in.
-        generate_new_cases : {"always", "attempt", "no"}, default "no"
-            How to generate new cases.
-        ordered_by_specified_features : bool, default False
-            If True order of generated feature values will match
-            the order of specified features.
-        num_cases_to_generate : int, default 1
-            The number of cases to generate.
-        new_case_threshold : {"min", "max", "most_similar"}, optional
-            Distance to determine the privacy cutoff. If None,
-            will default to "min".
-
-        Returns
-        -------
-        dict
-            The react result including audit details.
-        int
-            The request payload size.
-        int
-            The result payload size.
-        """
-        return self._execute_sized(trainee_id, "batch_react", {
-            "context_features": context_features,
-            "context_values": context_values,
-            "action_features": action_features,
-            "action_values": action_values,
-            "derived_context_features": derived_context_features,
-            "derived_action_features": derived_action_features,
-            "details": details,
-            "exclude_novel_nominals_from_uniqueness_check": exclude_novel_nominals_from_uniqueness_check,
-            "extra_features": extra_features,
-            "case_indices": case_indices,
-            "allow_nulls": allow_nulls,
-            "input_is_substituted": input_is_substituted,
-            "substitute_output": substitute_output,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-            "leave_case_out": leave_case_out,
-            "num_cases_to_generate": num_cases_to_generate,
-            "use_regional_model_residuals": use_regional_model_residuals,
-            "desired_conviction": desired_conviction,
-            "feature_bounds_map": feature_bounds_map,
-            "generate_new_cases": generate_new_cases,
-            "ordered_by_specified_features": ordered_by_specified_features,
-            "post_process_features": post_process_features,
-            "post_process_values": post_process_values,
-            "preserve_feature_values": preserve_feature_values,
-            "new_case_threshold": new_case_threshold,
-            "into_series_store": into_series_store,
-        })
-
     def batch_react_series(
         self,
         trainee_id: str,
